move from AutonomousUseCase to SimpleStrategy
